Introduce setting multi-lang helper
